ml_pipeline/requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
supabase
python-dotenv
requests
